Move manifest to separate class
