+++ src/logic.py
@@ -2,25 +2,11 @@ import os
 import pathlib
 import shutil
 import tempfile
-import re
-from dataclasses import dataclass
-from typing import Any
 
 from depot_downloader_helper import DepotDownloaderHelper
 from web_helper import WebHelper
+import manifest
 import utils
-
-
-@dataclass
-class Manifest():
-    depot: int
-    id: int
-    date: Any
-    num_files: int
-    num_chunks: int
-    size_disk: int
-    size_compressed: int
-    files: list[str]
 
 
 class Logic:
@@ -316,8 +302,8 @@ class Logic:
         self._download_manifest(username, depot_id, target_manifest_id)
 
         # Read manifest files
-        current_manifest = self._read_manifest(self.manifest_dir / f"manifest_{depot_id}_{current_manifest_id}.txt")
-        target_manifest = self._read_manifest(self.manifest_dir / f"manifest_{depot_id}_{target_manifest_id}.txt")
+        current_manifest = manifest.read_manifest(self.manifest_dir / f"manifest_{depot_id}_{current_manifest_id}.txt")
+        target_manifest = manifest.read_manifest(self.manifest_dir / f"manifest_{depot_id}_{target_manifest_id}.txt")
 
         # Initialize file sets
         current_set = set(current_manifest.files)
@@ -359,87 +345,9 @@ class Logic:
         self._download_manifest(username, depot_id, manifest_id)
 
         # Read manifest files
-        current_manifest = self._read_manifest(self.manifest_dir / f"manifest_{depot_id}_{manifest_id}.txt")
+        current_manifest = manifest.read_manifest(self.manifest_dir / f"manifest_{depot_id}_{manifest_id}.txt")
 
         return current_manifest.files
-
-    def _read_manifest(self, file: pathlib.Path) -> Manifest:
-        """Parse a given manifest file and return a manifest object.
-
-        Args:
-            file (pathlib.Path): Path to the manifest file
-
-        Returns:
-            Manifest: The parsed manifest object
-        """
-        def expectMatch(pattern: str, line: str) -> re.Match[str]:
-            match = re.match(pattern, line)
-
-            if match is None:
-                raise ValueError(f"Could not match patter '{pattern}' against line '{line}'")
-
-            return match
-
-        with open(file, "r") as f:
-            depot = 0
-            id = 0
-            date = 0
-            num_files = 0
-            num_chunks = 0
-            size_disk = 0
-            size_compressed = 0
-            files = []
-
-            try:
-                # First line contains depot id
-                line = f.readline()
-                depot = int(expectMatch(r".* (\d+)", line).groups()[0])
-
-                # Second lines is empty
-                # Third contains manifest id and date
-                line = f.readline()
-                line = f.readline()
-                groups = expectMatch(r".* : (\d+) \/ (.+)", line).groups()
-                id = int(groups[0])
-                date = str(groups[1])    # (Temporary) workaround since date isn't used anyways.
-                # Date format seems to be localized... @TODO find a way to universally parse datestring
-                # date = time.mktime(time.strptime(groups[1], "%d.%m.%Y %H:%M:%S"))
-
-                # Fourth line contains number of files
-                line = f.readline()
-                groups = expectMatch(r".* : (\d+)", line).groups()
-                num_files = int(groups[0])
-
-                # Fifth line contains number of chunks
-                line = f.readline()
-                groups = expectMatch(r".* : (\d+)", line).groups()
-                num_chunks = int(groups[0])
-
-                # Sixth line contains size on disk
-                line = f.readline()
-                groups = expectMatch(r".* : (\d+)", line).groups()
-                size_disk = int(groups[0])
-
-                # Seventh line contains size compressed
-                line = f.readline()
-                groups = expectMatch(r".* : (\d+)", line).groups()
-                size_compressed = int(groups[0])
-
-                # Eighth line is empty
-                # Ninth line is empty
-                # Tenth line contains headers
-                # Eleventh line until EOF contains one file per line
-                line = f.readline()
-                line = f.readline()
-                line = f.readline()
-                while line := f.readline():
-                    # Extract file hash and name
-                    match = expectMatch(r"\s+\d+\s+\d+\s+(?P<hash>.{40})\s+\d+\s+(?P<name>.+)", line)
-                    files.append((match.group("name"), match.group("hash")))
-
-                return Manifest(depot, id, date, num_files, num_chunks, size_disk, size_compressed, files)
-            except ValueError:
-                raise
 
     def _get_game_version(self) -> int:
         """Retrieve the game version from the executable file. Requires game directory to be set.
